Consider x-correlation-id as well
The library currently considers the header fields `x-correlationid` and `x-vcap-request-id` as carrying correlation IDs. This change would also make it aware of the field `x-correlation-id`, which is widely used in the web, today.

The settings would introduce `x-correlation-id` as an alias, sub-ordinating it to the original field `x-correlationid`. Meaning, for example, the library will use `x-correlationid` > `x-correation-id` > `x-vcap-request-id`, in this order of preference. The field would not be propagated back to the request sender, although this can probably be discussed.

This PR is an incomplete suggestion, to make the dimension of the changes to the production code clear. It does not adjust the unit alongside, so any automatic PR voters will/should reject this change.