Refactor response time measurement in RequestRecord
Removes the SimpleDateFormat in favour of java.time.Instant. This also removes
the `synchronized` block required to protect SDF. Tests were improved as well.
In total, this improves cpu usage during request handling. It is hard to give definite
numbers on the improvement, since the synchronized statement could have lead to
congestion.