Add logging of x-forwarded-* and x-custom-host http headers
Users can choose to log the following for http headers in the request logs:

* x-forwarded-for
* x-forwarded-host
* x-forwarded-proto
* x-custom-host

This requires the environment variable `LOG_SENSITIVE_CONNECTION_DATA` to be set to `true`.
If there are no headers, no fields will be added to the request logs. If there are headers, but
`LOG_SENSITIVE_CONNECTION_DATA` is set to `false`(default) the values will be logged as 'redacted'.