Suppress Default Values from MDC
The library can add fields from the MDC to each generated log messages.
Certain fields are always created, even if the value is unknown. This changes
avoids those fields in the output. This reduces message size and avoids
unnecessary fields in the target service.
The old behaviour is still available by an extra configuration in the logback.xml
or log4j2.xml.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>