Provide Dynatrace Exporter
Adds a DynatraceMetricsExporter to the provided exporters. This exporter will parse CF
bindings for a Dynatrace service binding. If found, it will export metrics to the API endpoint.
For this to work, the API token field in the service credentials needs to be provided with
`otel.javaagent.extension.sap.cf.binding.dynatrace.metrics.token-name`.

Several refacotrings align Dynatrace and Cloud Logging exporters.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>