PoC for TenantId Instrumentation
The sample application can set a tenantId to the span attribute.
It will be collected by the agent extension and transferred to all
log messages. This PoC can be enhanced to to span exports as
well. There are other options to achieve a similar result, but they
usually require a larger shared interface in application and agent.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>