Prove Key-Value-Pairs Support for Log4j2
SLF4J2 provides a fluent API, that allows to add key-value-pairs to the event context.
In Log4j2, the framework itself adds those fields to the context data. No additional
implementation is required, just a test to show the feature.

This commit contains a fix for a wrong package name.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>