Detect Bindings to SAP Application Logging Service
This library generates a special format for custom fields, that is
targeting SAP Application Logging Service. Since the library can
also be used with other logging services most prominently SAP
Cloud Logging, this special format needs to be disabled. Until now
this was possible by not declaring the custom fields or also
configuring them as retained fields.

This change introduces scanning of the CF service bindings for
SAP Application Logging Service. It will no longer create the
special format if no binding to that service is found. It is possible
to force the library to always emit the old format with environment
variable `LOG_GENERATE_APPLICATION_LOGGING_CUSTOM_FIELDS`
set to `true`.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>