Remove io.pivotal.cfenv
To avoid version conflicts, this dependency was removed.
The logic was rewritten to only use the JsonParser provided
by Jackson from the OTel Java Agent.

The resource provider is now built on the community version
of the CF resources and can even emit the attributes following
the official semantic conventions by configuration.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>