Fix Test Failures from ALS auto-detection
The new auto-detection for SAP Application Logging service for SAP BTP
introduced test regressions in the backend integrations. This change
introduces the ability to override the detection by system property and sets
the override for the existing tests. Furthermore a bug was fixed that initialised
a logger during SLF4j initiallization phase.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>